+++ reverse/test_sdk_examples.py
@@ -15,6 +15,7 @@ class SdkExamplesTest(unittest.TestCase):
     PREBUILT_EXAMPLES = [
         ("example/basic/hello_world/HelloWorld.bda", "HelloWorld", 9),
         ("example/filesystem/fs_write/FsWrite.bda", "FsWrite", 9),
+        ("example/filesystem/fs_flush/FsFlush.bda", "FsFlush", 9),
         ("example/input/key_polling/KeyInput.bda", "KeyInput", 9),
         ("example/input/touch_crosshair/TouchCrosshair.bda", "TouchXY", 9),
         ("example/graphics/primitives/GraphicsPrimitives.bda", "Graphics", 9),
@@ -135,6 +136,11 @@ class SdkExamplesTest(unittest.TestCase):
 
     def test_fs_write_example_builds(self) -> None:
         self.build_and_validate("example/filesystem/fs_write/fs_write_demo.c", "FsWrite")
+
+    def test_fs_flush_example_builds(self) -> None:
+        self.build_and_validate(
+            "example/filesystem/fs_flush/fs_flush_power_demo.c", "FsFlush"
+        )
 
     def test_key_msgbox_example_builds(self) -> None:
         self.build_and_validate("example/input/key_polling/key_msgbox_demo.c", "KeyInput")